# api/openapi.yml
openapi: 3.0.3
info:
  title: BookUp API
  description: >
    API del MVP de BookUp: catálogo de libros, sedes de biblioteca, reservas de
    ejemplares físicos y usuarios. Generado a partir del código en
    `api/app/controllers/` (routers FastAPI) y `api/app/controllers/schemas.py`.
  version: "0.1.0"
servers:
  - url: http://localhost:8000
    description: Desarrollo local
...
paths:
...
  /books:
    get:
      tags: [catalog]
      summary: Listar todos los libros del catálogo
      operationId: listBooks
      responses:
        "200":
          description: Lista de libros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookOut"
    post:
      tags: [catalog]
      summary: Crear un libro en el catálogo
      description: >
        Requiere `librarian` o `sysadmin`. El `isbn` tiene que ser un ISBN-13
        válido (13 dígitos y dígito verificador correcto), si no da 422.
        `author_ids`/`genre_ids` referencian autores y géneros ya existentes:
        si alguno no existe, 404.
      operationId: createBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookCreate"
      responses:
        "201":
          description: Libro creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /books/{isbn}:
    get:
      tags: [catalog]
      summary: Obtener un libro por ISBN
      operationId: getBook
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      responses:
        "200":
          description: Libro encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [catalog]
      summary: Actualizar un libro
      description: >
        Requiere `librarian` o `sysadmin`. Actualización parcial; solo se
        aplican los campos enviados. Enviar `author_ids`/`genre_ids`
        **reemplaza** la lista completa (mandar `[]` la vacía); omitirlos la
        deja como estaba.
      operationId: updateBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Libro actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [catalog]
      summary: Eliminar un libro del catálogo
      description: >
        Requiere `librarian` o `sysadmin`. Falla con 409 si el libro todavía
        tiene ejemplares físicos en alguna sede.
      operationId: deleteBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IsbnPath"
      responses:
        "204":
          description: Libro eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /physical-books:
    get:
      tags: [physical_books]
      summary: Listar ejemplares físicos
      operationId: listPhysicalBooks
      parameters:
        - name: isbn
          in: query
          required: false
          schema:
            type: string
        - name: library_id
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/PhysicalBookStatus"
      responses:
        "200":
          description: Lista de ejemplares
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhysicalBookOut"
        "422":
          $ref: "#/components/responses/ValidationError"
    post:
      tags: [physical_books]
      summary: Dar de alta un ejemplar en una sede
      description: >
        Requiere `sysadmin`, o el `librarian` de **esa** sede. El ejemplar nace
        en estado `available`. 404 si el libro o la sede no existen.
      operationId: createPhysicalBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhysicalBookCreate"
      responses:
        "201":
          description: Ejemplar creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalBookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /physical-books/{physical_book_id}:
    get:
      tags: [physical_books]
      summary: Obtener un ejemplar por id
      operationId: getPhysicalBook
      parameters:
        - $ref: "#/components/parameters/PhysicalBookIdPath"
      responses:
        "200":
          description: Ejemplar encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalBookOut"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [physical_books]
      summary: Dar de baja un ejemplar
      description: >
        Requiere `sysadmin`, o el `librarian` de la sede que lo tiene. Falla con
        409 si el ejemplar todavía tiene reservas asociadas.
      operationId: deletePhysicalBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PhysicalBookIdPath"
      responses:
        "204":
          description: Ejemplar eliminado
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /physical-books/{physical_book_id}/status:
    patch:
      tags: [physical_books]
      summary: Cambiar el estado de un ejemplar
      description: >
        Requiere `sysadmin`, o el `librarian` de la sede que lo tiene. Solo
        admite las transiciones manuales `available` ↔ `lost`: pedir `reserved`
        o `loaned` da 409 porque esos estados los maneja el flujo de reservas
        (`POST /reservations`, `PATCH /reservations/{id}/pickup`), y un ejemplar
        que hoy está `reserved`/`loaned` tampoco se puede tocar por acá — hay
        que resolver su reserva primero.
      operationId: updatePhysicalBookStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PhysicalBookIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhysicalBookStatusUpdate"
      responses:
        "200":
          description: Ejemplar actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalBookOut"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT emitido por `POST /auth/login`, enviado en el header
        `Authorization` con el prefijo `Bearer`.
  parameters:
    IsbnPath:
      name: isbn
      in: path
      required: true
      schema:
        type: string
...
    PhysicalBookIdPath:
      name: physical_book_id
      in: path
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Unauthorized:
      description: Falta el token, está vencido o es inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Forbidden:
      description: El usuario está autenticado pero no tiene permiso para esta operación
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    Conflict:
      description: La operación entra en conflicto con el estado actual del recurso
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDetail"
    ValidationError:
      description: Error de validación del payload/parámetros (FastAPI/Pydantic)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPValidationError"
  schemas:
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
      example:
        detail: "Book 9780000000000 not found"
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string
...
    AuthorOut:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    GenreOut:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    BookBase:
      type: object
      required: [isbn, title, language]
      properties:
        isbn:
          type: string
        title:
          type: string
        language:
          type: string
        pages:
          type: integer
          nullable: true
        synopsis:
          type: string
          nullable: true
    BookOut:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            authors:
              type: array
              items:
                $ref: "#/components/schemas/AuthorOut"
              default: []
            genres:
              type: array
              items:
                $ref: "#/components/schemas/GenreOut"
              default: []
    BookCreate:
      allOf:
        - $ref: "#/components/schemas/BookBase"
        - type: object
          properties:
            isbn:
              type: string
              description: ISBN-13 válido (13 dígitos y dígito verificador correcto).
              pattern: "^[0-9]{13}$"
            author_ids:
              type: array
              items:
                type: integer
              default: []
            genre_ids:
              type: array
              items:
                type: integer
              default: []
    BookUpdate:
      type: object
      description: >
        Actualización parcial. `author_ids`/`genre_ids` reemplazan la lista
        completa cuando se envían.
      properties:
        title:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        pages:
          type: integer
          nullable: true
        synopsis:
          type: string
          nullable: true
        author_ids:
          type: array
          items:
            type: integer
          nullable: true
        genre_ids:
          type: array
          items:
            type: integer
          nullable: true
    PhysicalBookStatus:
      type: string
      enum: [available, reserved, loaned, lost]
    PhysicalBookOut:
      type: object
      required: [id, isbn, library_id, status]
      properties:
        id:
          type: integer
        isbn:
          type: string
        library_id:
          type: integer
        status:
          $ref: "#/components/schemas/PhysicalBookStatus"
    PhysicalBookCreate:
      type: object
      required: [isbn, library_id]
      properties:
        isbn:
          type: string
          description: ISBN-13 válido de un libro ya existente en el catálogo.
          pattern: "^[0-9]{13}$"
        library_id:
          type: integer
    PhysicalBookStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/PhysicalBookStatus"
...